Move AwsProvider plugin into aws directory
