+++ lib/plugins/aws/deploy/compile/events/apiGateway/tests/index.js
@@ -3,7 +3,7 @@
 const expect = require('chai').expect;
 const sinon = require('sinon');
 const BbPromise = require('bluebird');
-const AwsProvider = require('../../../../../../awsProvider/awsProvider');
+const AwsProvider = require('../../../../../provider/awsProvider');
 const AwsCompileApigEvents = require('../index');
 const Serverless = require('../../../../../../../Serverless');
