+++ lib/plugins/aws/deploy/compile/events/s3/tests/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../awsProvider/awsProvider');
+const AwsProvider = require('../../../../../provider/awsProvider');
 const AwsCompileS3Events = require('../index');
 const Serverless = require('../../../../../../../Serverless');
