+++ lib/plugins/aws/deploy/compile/events/sns/tests/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../awsProvider/awsProvider');
+const AwsProvider = require('../../../../../provider/awsProvider');
 const AwsCompileSNSEvents = require('../index');
 const Serverless = require('../../../../../../../Serverless');
