+++ lib/plugins/aws/deploy/compile/events/stream/tests/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../awsProvider/awsProvider');
+const AwsProvider = require('../../../../../provider/awsProvider');
 const AwsCompileStreamEvents = require('../index');
 const Serverless = require('../../../../../../../Serverless');
