+++ lib/plugins/aws/deploy/compile/functions/tests/index.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../awsProvider/awsProvider');
+const AwsProvider = require('../../../../provider/awsProvider');
 const AwsCompileFunctions = require('../index');
 const Serverless = require('../../../../../../Serverless');
