+++ lib/plugins/aws/deploy/tests/cleanupS3Bucket.js
@@ -3,7 +3,7 @@
 const sinon = require('sinon');
 const BbPromise = require('bluebird');
 const expect = require('chai').expect;
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsDeploy = require('../index');
 const Serverless = require('../../../../Serverless');
