+++ lib/plugins/aws/deploy/tests/configureStack.js
@@ -4,7 +4,7 @@ const sinon = require('sinon');
 const BbPromise = require('bluebird');
 const path = require('path');
 const expect = require('chai').expect;
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const Serverless = require('../../../../Serverless');
 const validate = require('../../lib/validate');
 const configureStack = require('../lib/configureStack');
