+++ lib/plugins/aws/deploy/tests/createStack.js
@@ -4,7 +4,7 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 const path = require('path');
 const BbPromise = require('bluebird');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsDeploy = require('../index');
 const Serverless = require('../../../../Serverless');
 const testUtils = require('../../../../../tests/utils');
