+++ lib/plugins/aws/deploy/tests/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsDeploy = require('../index');
 const Serverless = require('../../../../Serverless');
 const expect = require('chai').expect;
