+++ lib/plugins/aws/deployFunction/tests/index.js
@@ -5,7 +5,7 @@ const sinon = require('sinon');
 const path = require('path');
 const fs = require('fs');
 const Package = require('../../../package');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsDeployFunction = require('../index');
 const Serverless = require('../../../../Serverless');
 const BbPromise = require('bluebird');
