+++ lib/plugins/aws/info/tests/index.js
@@ -3,7 +3,7 @@
 const expect = require('chai').expect;
 const sinon = require('sinon');
 const AwsInfo = require('../');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const Serverless = require('../../../../Serverless');
 const CLI = require('../../../../classes/CLI');
 const BbPromise = require('bluebird');
