+++ lib/plugins/aws/invoke/tests/index.js
@@ -4,7 +4,7 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 const path = require('path');
 const AwsInvoke = require('../');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const Serverless = require('../../../../Serverless');
 const BbPromise = require('bluebird');
 const testUtils = require('../../../../../tests/utils');
