+++ lib/plugins/aws/logs/tests/index.js
@@ -2,7 +2,7 @@
 
 const expect = require('chai').expect;
 const sinon = require('sinon');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsLogs = require('../');
 const Serverless = require('../../../../Serverless');
 const BbPromise = require('bluebird');
