+++ lib/plugins/aws/provider/awsProvider.js
