+++ lib/plugins/aws/provider/awsProvider.test.js
@@ -3,7 +3,7 @@
 const sinon = require('sinon');
 const BbPromise = require('bluebird');
 const expect = require('chai').expect;
-const Serverless = require('../../Serverless');
+const Serverless = require('../../../Serverless');
 const AwsProvider = require('./awsProvider');
 const proxyquire = require('proxyquire');
