+++ lib/plugins/aws/remove/tests/index.js
@@ -3,7 +3,7 @@
 const expect = require('chai').expect;
 const BbPromise = require('bluebird');
 const sinon = require('sinon');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsRemove = require('../index');
 const Serverless = require('../../../../Serverless');
