+++ lib/plugins/aws/remove/tests/stack.js
@@ -2,7 +2,7 @@
 
 const expect = require('chai').expect;
 const sinon = require('sinon');
-const AwsProvider = require('../../../awsProvider/awsProvider');
+const AwsProvider = require('../../provider/awsProvider');
 const AwsRemove = require('../index');
 const Serverless = require('../../../../Serverless');
 const BbPromise = require('bluebird');
