+++ lib/plugins/aws/tests/monitorStack.js
@@ -4,7 +4,7 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 const BbPromise = require('bluebird');
 const Serverless = require('../../../Serverless');
-const AwsProvider = require('../../awsProvider/awsProvider');
+const AwsProvider = require('../provider/awsProvider');
 const CLI = require('../../../classes/CLI');
 const monitorStack = require('../lib/monitorStack');
